requirements.txt
discord
asyncpraw
aiohttp
asyncio
helper
discord.ext
python-dotenv
gitpython